feat(map): refactor map 3D mode logic and extract to custom hook
